backend/requirements-prod.txt
fastapi==0.110.1
uvicorn==0.29.0
python-dotenv==1.1.0
pydantic[email]==2.7.0
requests==2.31.0
bcrypt==4.1.3
PyJWT==2.8.0
yfinance==0.2.37
apscheduler==3.10.4
resend==0.8.0
